ecommerce/ai-agent/requirements.txt
langchain
langgraph
langchain-google-genai
langchain-community
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
PyJWT
